requirements.txt
python-telegram-bot==20.7
pymongo==4.6.1
motor==3.3.2
python-dotenv==1.0.0
yt-dlp==2023.12.30
aiohttp==3.9.1
aiofiles==23.2.0
pillow
opencv-python==4.8.1.78
asyncio-throttle==1.0.2
validators==0.22.0
python-magic-bin==0.4.14
requests==2.31.0
beautifulsoup4==4.12.2
